+++ test/selectors/optimizers/simple-test.js
@@ -72,6 +72,10 @@ vows.describe(SimpleOptimizer)
       '+html': [
         '*+html .foo{display:inline}',
         null
+      ],
+      'adjacent nav': [
+        'div + nav{}',
+        [{ value: 'div+nav' }]
       ]
     })
   )
@@ -102,6 +106,18 @@ vows.describe(SimpleOptimizer)
         [{ value: '*+html .foo' }, { value: '.bar' }]
       ]
     }, { compatibility: 'ie7' })
+  )
+  .addBatch(
+    selectorContext('+adjacentSpace', {
+      'with whitespace': [
+        'div + nav{}',
+        [{ value: 'div+ nav' }]
+      ],
+      'without whitespace': [
+        'div+nav{}',
+        [{ value: 'div+ nav' }]
+      ]
+    }, { compatibility: { selectors: { adjacentSpace: true } } })
   )
   .addBatch(
     propertyContext('@background', {
@@ -490,6 +506,14 @@ vows.describe(SimpleOptimizer)
         ['text-shadow:rgba(255,1,1,.5)1px']
       ]
     })
+  )
+  .addBatch(
+    propertyContext('whitespace in compatibility mode', {
+      'stripped spaces': [
+        'div{text-shadow:rgba(255,1,1,.5) 1px}',
+        ['text-shadow:rgba(255,1,1,.5) 1px']
+      ]
+    }, { compatibility: 'ie8' })
   )
   .addBatch(
     propertyContext('whitespace in compatibility mode', {